fix: namespace and module structure
